// htdocs/rescueman/src/app/shared/models/person.ts
export class Person {
    id: number;
    name: string;
    email: string;
    phone: string;
    home_phone: string;
    business_phone: string;
    street_line_1: string;
    street_line_2: string;
    city: string;
    state_prov: string;
    postal_code: string;
    roles: string;
    volunteer: string;
    foster: string;
    can_foster: string;
    home_inspected: boolean;
    home_inspected_by: number;
    adopter_contract: boolean;
    created_by: number;
    updated_by: number;
    statecode: string;
    created_at: Date;
    updated_at: Date;
}